Move Game type into /src/types/apiData.d.ts and create separate types for game responses, which include id, created_at, and updated_at, and requests, which don't include these values
